Modularize models.py with circular import hacks
