+++ server/models/model/Grant.py
@@ -1,0 +1,28 @@
+from server.models.db import db, Model, StringList
+
+class Grant(Model):
+    id = db.Column(db.Integer, primary_key=True)
+
+    user_id = db.Column(
+        db.Integer, db.ForeignKey('user.id', ondelete='CASCADE')
+    )
+    user = db.relationship('User')
+
+    client_id = db.Column(
+        db.String(40), db.ForeignKey('client.client_id'),
+        nullable=False,
+    )
+    client = db.relationship('Client')
+
+    code = db.Column(db.String(255), index=True, nullable=False)
+
+    redirect_uri = db.Column(db.String(255))
+    expires = db.Column(db.DateTime)
+
+    scopes = db.Column(StringList)
+
+    def delete(self):
+        db.session.delete(self)
+        db.session.commit()
+        return self
+
