+++ server/models/model/Token.py
@@ -1,0 +1,27 @@
+from server.models.db import db, Model, StringList
+
+class Token(Model):
+    id = db.Column(db.Integer, primary_key=True)
+    client_id = db.Column(
+        db.String(40), db.ForeignKey('client.client_id'),
+        nullable=False,
+    )
+    client = db.relationship('Client')
+
+    user_id = db.Column(
+        db.Integer, db.ForeignKey('user.id')
+    )
+    user = db.relationship('User')
+
+    # currently only bearer is supported
+    token_type = db.Column(db.String(40))
+
+    access_token = db.Column(db.String(255), unique=True)
+    refresh_token = db.Column(db.String(255), unique=True)
+    expires = db.Column(db.DateTime)
+    scopes = db.Column(StringList)
+
+    def delete(self):
+        db.session.delete(self)
+        db.session.commit()
+        return self
